Change source code files directories
