+++ kernel/drivers/vga/framebuffer.c
@@ -1,6 +1,6 @@
 #include <init/multiboot.h>
+#include <kernel/vga.h>
 #include <stdint.h>
-#include <vga/vga.h>
 
 /* Framebuffer Info */
 static uint32_t framebuffer_width;
