--- kernel/drivers/vga/framebuffer.h
@@ -1,7 +1,0 @@
-#ifndef _DRIVERS_VGA_FRAMEBUFFER_H
-#define _DRIVERS_VGA_FRAMEBUFFER_H
-
-void put_pixel(int pos_x, int pos_y, unsigned char color);
-void init_framebuffer();
-
-#endif //_DRIVERS_VGA_FRAMEBUFFER_H
