+++ kernel/fs/tarfs/tarfs.c
@@ -1,5 +1,5 @@
-#include "tarfs.h"
 #include <kernel/kheap.h>
+#include <kernel/tarfs.h>
 #include <stdint.h>
 
 struct tarfs_header *headers[32];
