+++ kernel/include/kernel/framebuffer.h
@@ -1,0 +1,7 @@
+#ifndef _KERNEL_FRAMEBUFFER_H
+#define _KERNEL_FRAMEBUFFER_H
+
+void put_pixel(int pos_x, int pos_y, unsigned char color);
+void init_framebuffer();
+
+#endif //_KERNEL_FRAMEBUFFER_H
