+++ kernel/include/kernel/pcspkr.h
@@ -1,5 +1,5 @@
-#ifndef _DRIVERS_SOUND_PCSPKR_H
-#define _DRIVERS_SOUND_PCSPKR_H
+#ifndef _KERNEL_PCSPKR_H
+#define _KERNEL_PCSPKR_H
 
 #include <stdint.h>
 
@@ -8,4 +8,4 @@ void play_sound_pcspkr(uint32_t frequency);
 void stop_sound_pcspkr();
 int init_pcspkr();
 
-#endif //_DRIVERS_SOUND_PCSPKR_H
+#endif //_KERNEL_PCSPKR_H
