+++ kernel/include/kernel/pit.h
@@ -1,8 +1,8 @@
-#ifndef _DRIVERS_PIT_H
-#define _DRIVERS_PIT_H
+#ifndef _KERNEL_PIT_H
+#define _KERNEL_PIT_H
 
 void handler_pit();
 void sleep_pit(int ticks);
 void init_pit(int hz);
 
-#endif //_DRIVERS_PIT_H
+#endif //_KERNEL_PIT_H
