+++ kernel/include/kernel/serial.h
@@ -1,5 +1,5 @@
-#ifndef _DRIVERS_SERIAL_H
-#define _DRIVERS_SERIAL_H
+#ifndef _KERNEL_SERIAL_H
+#define _KERNEL_SERIAL_H
 
 void init_serial();
 int received_serial();
@@ -8,4 +8,4 @@ int transmit_empty();
 void writechar_serial(char c);
 void writestr_serial(const char *str);
 
-#endif //_DRIVERS_SERIAL_H
+#endif //_KERNEL_SERIAL_H
