+++ kernel/init/multiboot.h
