+++ kernel/initrd/file.txt
