+++ kernel/initrd/file2.txt
