+++ kernel/initrd/initrd.c
