+++ kernel/kernel/printm/printm.c
@@ -1,6 +1,6 @@
+#include <kernel/vga.h>
 #include <stdarg.h>
 #include <stdio.h>
-#include <vga/vga.h>
 
 #define PRINTM_BUF_SIZE 1024
