+++ user/commands/echo.c
@@ -1,6 +1,6 @@
 #include "commands.h"
+#include <kernel/vga.h>
 #include <string.h>
-#include <vga/vga.h>
 
 int echo(char args[]) {
   if (strlen(args) == 0) {
