+++ user/commands/uname.c
@@ -1,7 +1,7 @@
 #include "commands.h"
 #include <generated/config.h>
+#include <kernel/vga.h>
 #include <string.h>
-#include <vga/vga.h>
 
 int uname(char args[]) {
   if (strcmp(args, "-s") == 0) {
