+++ user/commands/version.c
@@ -1,5 +1,5 @@
 #include "commands.h"
-#include <vga/vga.h>
+#include <kernel/vga.h>
 
 void version() {
   writestr("Version 0.11-rc2\n");
